requirements.txt
grpcio-tools==1.51.1
pydantic==2.8.2
pythonnet==3.0.1
requests==2.29.0 
types-requests==2.28.11.17
ruff==0.0.256
mypy==1.11.2
mypy-protobuf==3.4.0
types-protobuf==4.22.0.0
types_pywin32==305.0.0.11
types_pyserial==3.5.0.7
grpc-stubs==1.24.12.1
networkx==3.1
networkx-stubs==0.0.1
types-setuptools==68.0.0.2
slackclient==2.9.4
python-dotenv==1.0.0
pytest==7.4.3
msl-loadlib==0.10.0
pywin32==304; platform_system == "Windows"
typing-inspect==0.8.0
typing_extensions==4.8.0
pyserial==3.5
slack-sdk==3.21.3
slackclient==2.9.4
cython==3.0.11
pillow==11.0.0